8popitka

+++ src/exercise1.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 int main(int argc, char** argv) {
-    float array[10];
-    float sum = 0;
+    double array[10];
+    double sum = 0;
     for(int i = 0; i < 10; ++i) {
-        scanf("%f", &array[i]);
+        scanf("%lf", &array[i]);
         sum +=array[i];
  }
-    float ot = (sum/10);
+    double ot = (sum/10);
     if (ot>0){
         ot = ((sum/10)*100+0.5)/100;
     }
